Enhance TypeScript configuration and add shared module aliasing for better code organization. Update electron.vite.config.ts, tsconfig.json, tsconfig.node.json, and tsconfig.web.json to include '@shared' alias. Refactor VersementDetailPage to utilize shared schema for form validation.
