+++ tsconfig.json
@@ -8,6 +8,9 @@
     "paths": {
       "@/*": [
         "src/renderer/src/*"
+      ],
+      "@shared/*": [
+        "src/shared/*"
       ]
     }
   }
