+++ tsconfig.node.json
@@ -1,6 +1,6 @@
 {
   "extends": "@electron-toolkit/tsconfig/tsconfig.node.json",
-  "include": ["electron.vite.config.*", "src/main/**/*", "src/preload/**/*", "type.ts"],
+  "include": ["electron.vite.config.*", "src/main/**/*", "src/preload/**/*", "type.ts","src/shared/**/*"],
   "compilerOptions": {
     "composite": true,
     "types": ["electron-vite/node"]
