+++ tsconfig.web.json
@@ -5,6 +5,7 @@
     "src/renderer/src/**/*",
     "src/renderer/src/**/*.tsx",
     "src/preload/*.d.ts",
+    "src/shared/**/*",
     "type.ts"
   ],
   "compilerOptions": {
@@ -14,6 +15,9 @@
     "paths": {
       "@/*": [
         "src/renderer/src/*"
+      ],
+      "@shared/*": [
+        "src/shared/*"
       ]
     }
   }
